+++ internal/query/conn/options.go
@@ -1,0 +1,9 @@
+package conn
+
+type Option func(c *Conn)
+
+func WithOnClose(onClose func()) Option {
+	return func(c *Conn) {
+		c.onClose = append(c.onClose, onClose)
+	}
+}