+++ internal/table/session.go
@@ -46,15 +46,15 @@ import (
 // Note that after session is no longer needed it should be destroyed by
 // Close() call.
 type session struct {
-	onClose      []func(s *session)
-	id           string
-	tableService Ydb_Table_V1.TableServiceClient
-	status       table.SessionStatus
-	config       *config.Config
-	lastUsage    atomic.Int64
-	statusMtx    sync.RWMutex
-	closeOnce    sync.Once
-	nodeID       atomic.Uint32
+	onClose   []func(s *session)
+	id        string
+	client    Ydb_Table_V1.TableServiceClient
+	status    table.SessionStatus
+	config    *config.Config
+	lastUsage atomic.Int64
+	statusMtx sync.RWMutex
+	closeOnce sync.Once
+	nodeID    atomic.Uint32
 }
 
 func (s *session) IsAlive() bool {
@@ -114,6 +114,28 @@ func (s *session) isClosed() bool {
 	return s.Status() == table.SessionClosed
 }
 
+func Session(id string, cc grpc.ClientConnInterface, config *config.Config) *session {
+	s := &session{
+		id:     id,
+		config: config,
+		status: table.SessionReady,
+	}
+
+	s.lastUsage.Store(time.Now().Unix())
+	s.client = Ydb_Table_V1.NewTableServiceClient(
+		conn.WithBeforeFunc(
+			conn.WithContextModifier(cc, func(ctx context.Context) context.Context {
+				return meta.WithTrailerCallback(balancerContext.WithNodeID(ctx, s.NodeID()), s.checkCloseHint)
+			}),
+			func() {
+				s.lastUsage.Store(time.Now().Unix())
+			},
+		),
+	)
+
+	return s
+}
+
 func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (
 	s *session, err error,
 ) {
@@ -123,11 +145,13 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config
 	defer func() {
 		onDone(s, err)
 	}()
+
 	var (
 		response *Ydb_Table.CreateSessionResponse
 		result   Ydb_Table.CreateSessionResult
 		c        = Ydb_Table_V1.NewTableServiceClient(cc)
 	)
+
 	response, err = c.CreateSession(ctx,
 		&Ydb_Table.CreateSessionRequest{
 			OperationParams: operation.Params(
@@ -141,30 +165,13 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
+
 	err = response.GetOperation().GetResult().UnmarshalTo(&result)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
 
-	s = &session{
-		id:     result.GetSessionId(),
-		config: config,
-		status: table.SessionReady,
-	}
-	s.lastUsage.Store(time.Now().Unix())
-
-	s.tableService = Ydb_Table_V1.NewTableServiceClient(
-		conn.WithBeforeFunc(
-			conn.WithContextModifier(cc, func(ctx context.Context) context.Context {
-				return meta.WithTrailerCallback(balancerContext.WithNodeID(ctx, s.NodeID()), s.checkCloseHint)
-			}),
-			func() {
-				s.lastUsage.Store(time.Now().Unix())
-			},
-		),
-	)
-
-	return s, nil
+	return Session(result.GetSessionId(), cc, config), nil
 }
 
 func (s *session) ID() string {
@@ -191,7 +198,7 @@ func (s *session) Close(ctx context.Context) (err error) {
 		}()
 
 		if time.Since(s.LastUsage()) < s.config.IdleThreshold() {
-			_, err = s.tableService.DeleteSession(ctx,
+			_, err = s.client.DeleteSession(ctx,
 				&Ydb_Table.DeleteSessionRequest{
 					SessionId: s.id,
 					OperationParams: operation.Params(ctx,
@@ -242,7 +249,7 @@ func (s *session) KeepAlive(ctx context.Context) (err error) {
 		onDone(err)
 	}()
 
-	resp, err := s.tableService.KeepAlive(ctx,
+	resp, err := s.client.KeepAlive(ctx,
 		&Ydb_Table.KeepAliveRequest{
 			SessionId: s.id,
 			OperationParams: operation.Params(
@@ -297,7 +304,7 @@ func (s *session) CreateTable(
 			opt.ApplyCreateTableOption((*options.CreateTableDesc)(&request), a)
 		}
 	}
-	_, err = s.tableService.CreateTable(ctx, &request)
+	_, err = s.client.CreateTable(ctx, &request)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}
@@ -311,8 +318,8 @@ func (s *session) DescribeTable(
 	path string,
 	opts ...options.DescribeTableOption,
 ) (desc options.Description, err error) {
-	request := initializeRequest(ctx, s, path, opts)
-	response, err := s.tableService.DescribeTable(ctx, request)
+	request := describeTableRequest(ctx, s, path, opts)
+	response, err := s.client.DescribeTable(ctx, request)
 	if err != nil {
 		return desc, xerrors.WithStackTrace(err)
 	}
@@ -343,7 +350,7 @@ func (s *session) DescribeTable(
 	return desc, nil
 }
 
-func initializeRequest(
+func describeTableRequest(
 	ctx context.Context,
 	s *session,
 	path string,
@@ -504,7 +511,7 @@ func (s *session) DropTable(
 			opt.ApplyDropTableOption((*options.DropTableDesc)(&request))
 		}
 	}
-	_, err = s.tableService.DropTable(ctx, &request)
+	_, err = s.client.DropTable(ctx, &request)
 
 	return xerrors.WithStackTrace(err)
 }
@@ -544,7 +551,7 @@ func (s *session) AlterTable(
 			opt.ApplyAlterTableOption((*options.AlterTableDesc)(&request), a)
 		}
 	}
-	_, err = s.tableService.AlterTable(ctx, &request)
+	_, err = s.client.AlterTable(ctx, &request)
 
 	return xerrors.WithStackTrace(err)
 }
@@ -571,7 +578,7 @@ func (s *session) CopyTable(
 			opt((*options.CopyTableDesc)(&request))
 		}
 	}
-	_, err = s.tableService.CopyTable(ctx, &request)
+	_, err = s.client.CopyTable(ctx, &request)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}
@@ -621,7 +628,7 @@ func (s *session) CopyTables(
 	ctx context.Context,
 	opts ...options.CopyTablesOption,
 ) (err error) {
-	err = copyTables(ctx, s.id, s.config.OperationTimeout(), s.config.OperationCancelAfter(), s.tableService, opts...)
+	err = copyTables(ctx, s.id, s.config.OperationTimeout(), s.config.OperationCancelAfter(), s.client, opts...)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}
@@ -671,7 +678,7 @@ func (s *session) RenameTables(
 	ctx context.Context,
 	opts ...options.RenameTablesOption,
 ) (err error) {
-	err = renameTables(ctx, s.id, s.config.OperationTimeout(), s.config.OperationCancelAfter(), s.tableService, opts...)
+	err = renameTables(ctx, s.id, s.config.OperationTimeout(), s.config.OperationCancelAfter(), s.client, opts...)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}
@@ -704,7 +711,7 @@ func (s *session) Explain(
 		}
 	}()
 
-	response, err = s.tableService.ExplainDataQuery(ctx,
+	response, err = s.client.ExplainDataQuery(ctx,
 		&Ydb_Table.ExplainDataQueryRequest{
 			SessionId: s.id,
 			YqlText:   query,
@@ -753,7 +760,7 @@ func (s *session) Prepare(ctx context.Context, queryText string) (_ table.Statem
 		}
 	}()
 
-	response, err = s.tableService.PrepareDataQuery(ctx,
+	response, err = s.client.PrepareDataQuery(ctx,
 		&Ydb_Table.PrepareDataQueryRequest{
 			SessionId: s.id,
 			YqlText:   queryText,
@@ -880,7 +887,7 @@ func (s *session) executeDataQuery(
 		response *Ydb_Table.ExecuteDataQueryResponse
 	)
 
-	response, err = s.tableService.ExecuteDataQuery(ctx, request, callOptions...)
+	response, err = s.client.ExecuteDataQuery(ctx, request, callOptions...)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -914,7 +921,7 @@ func (s *session) ExecuteSchemeQuery(
 			opt((*options.ExecuteSchemeQueryDesc)(&request))
 		}
 	}
-	_, err = s.tableService.ExecuteSchemeQuery(ctx, &request)
+	_, err = s.client.ExecuteSchemeQuery(ctx, &request)
 
 	return xerrors.WithStackTrace(err)
 }
@@ -938,7 +945,7 @@ func (s *session) DescribeTableOptions(ctx context.Context) (
 			operation.ModeSync,
 		),
 	}
-	response, err = s.tableService.DescribeTableOptions(ctx, &request)
+	response, err = s.client.DescribeTableOptions(ctx, &request)
 	if err != nil {
 		return desc, xerrors.WithStackTrace(err)
 	}
@@ -1091,7 +1098,7 @@ func (s *session) StreamReadTable(
 
 	ctx, cancel := xcontext.WithCancel(ctx)
 
-	stream, err = s.tableService.StreamReadTable(ctx, &request)
+	stream, err = s.client.StreamReadTable(ctx, &request)
 	if err != nil {
 		cancel()
 
@@ -1153,7 +1160,7 @@ func (s *session) ReadRows(
 		}
 	}
 
-	response, err = s.tableService.ReadRows(ctx, &request)
+	response, err = s.client.ReadRows(ctx, &request)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -1213,7 +1220,7 @@ func (s *session) StreamExecuteScanQuery(
 
 	ctx, cancel := xcontext.WithCancel(ctx)
 
-	stream, err = s.tableService.StreamExecuteScanQuery(ctx, &request, callOptions...)
+	stream, err = s.client.StreamExecuteScanQuery(ctx, &request, callOptions...)
 	if err != nil {
 		cancel()
 
@@ -1273,7 +1280,7 @@ func (s *session) BulkUpsert(ctx context.Context, table string, rows value.Value
 		}
 	}
 
-	_, err = s.tableService.BulkUpsert(ctx,
+	_, err = s.client.BulkUpsert(ctx,
 		&Ydb_Table.BulkUpsertRequest{
 			Table: table,
 			Rows:  value.ToYDB(rows, a),
@@ -1311,7 +1318,7 @@ func (s *session) BeginTransaction(
 		onDone(x, err)
 	}()
 
-	response, err = s.tableService.BeginTransaction(ctx,
+	response, err = s.client.BeginTransaction(ctx,
 		&Ydb_Table.BeginTransactionRequest{
 			SessionId:  s.id,
 			TxSettings: txSettings.Settings(),