+++ internal/table/transaction.go
@@ -172,7 +172,7 @@ func (tx *transaction) CommitTx(
 			}
 		}
 
-		response, err = tx.s.tableService.CommitTransaction(ctx, request)
+		response, err = tx.s.client.CommitTransaction(ctx, request)
 		if err != nil {
 			return nil, xerrors.WithStackTrace(err)
 		}
@@ -209,7 +209,7 @@ func (tx *transaction) Rollback(ctx context.Context) (err error) {
 	case txStateRollbacked:
 		return xerrors.WithStackTrace(errTxRollbackedEarly)
 	default:
-		_, err = tx.s.tableService.RollbackTransaction(ctx,
+		_, err = tx.s.client.RollbackTransaction(ctx,
 			&Ydb_Table.RollbackTransactionRequest{
 				SessionId: tx.s.id,
 				TxId:      tx.ID(),